+++ test/session_test/cookies_test.rb
@@ -98,6 +98,12 @@ module SessionTest
       def test_credentials
         session = UserSession.new
         session.credentials = { remember_me: true }
+        assert_equal({}, session.credentials)
+      end
+
+      def test_credentials_assignment
+        session = UserSession.new
+        session.credentials = { remember_me: true }
         assert_equal true, session.remember_me
       end